move createVehicleFootprints() to seperate files
