+++ control/autoware_lane_departure_checker/include/autoware/lane_departure_checker/utils.hpp
@@ -15,15 +15,22 @@
 #ifndef AUTOWARE__LANE_DEPARTURE_CHECKER__UTILS_HPP_
 #define AUTOWARE__LANE_DEPARTURE_CHECKER__UTILS_HPP_
 
+#include <autoware/universe_utils/geometry/boost_geometry.hpp>
+#include <autoware_vehicle_info_utils/vehicle_info.hpp>
+
 #include <autoware_planning_msgs/msg/trajectory.hpp>
 #include <autoware_planning_msgs/msg/trajectory_point.hpp>
+#include <geometry_msgs/msg/pose_with_covariance.hpp>
+#include <tier4_planning_msgs/msg/path_with_lane_id.hpp>
 
 #include <vector>
 
 namespace autoware::lane_departure_checker::utils
 {
+using autoware::universe_utils::LinearRing2d;
 using autoware_planning_msgs::msg::Trajectory;
 using autoware_planning_msgs::msg::TrajectoryPoint;
+using tier4_planning_msgs::msg::PathWithLaneId;
 using TrajectoryPoints = std::vector<TrajectoryPoint>;
 
 /**
@@ -42,6 +49,30 @@ TrajectoryPoints cutTrajectory(const TrajectoryPoints & trajectory, const double
  * @note this function assumes the input trajectory is sampled dense enough
  */
 TrajectoryPoints resampleTrajectory(const Trajectory & trajectory, const double interval);
+
+/**
+ * @brief create vehicle footprints along the trajectory with the given covariance and margin
+ * @param covariance vehicle pose with covariance
+ * @param trajectory trajectory along which the vehicle footprints are created
+ * @param vehicle_info vehicle information
+ * @param footprint_margin_scale scale of the footprint margin
+ * @return vehicle footprints along the trajectory
+ */
+std::vector<LinearRing2d> createVehicleFootprints(
+  const geometry_msgs::msg::PoseWithCovariance & covariance, const TrajectoryPoints & trajectory,
+  const autoware::vehicle_info_utils::VehicleInfo & vehicle_info,
+  const double footprint_margin_scale);
+
+/**
+ * @brief create vehicle footprints along the path with the given margin
+ * @param path path along which the vehicle footprints are created
+ * @param vehicle_info vehicle information
+ * @param footprint_extra_margin extra margin for the footprint
+ * @return vehicle footprints along the path
+ */
+std::vector<LinearRing2d> createVehicleFootprints(
+  const PathWithLaneId & path, const autoware::vehicle_info_utils::VehicleInfo & vehicle_info,
+  const double footprint_extra_margin);
 }  // namespace autoware::lane_departure_checker::utils
 
 #endif  // AUTOWARE__LANE_DEPARTURE_CHECKER__UTILS_HPP_
