+++ control/autoware_lane_departure_checker/src/lane_departure_checker_node/utils.cpp
@@ -18,6 +18,38 @@
 
 #include <boost/geometry.hpp>
 
+namespace
+{
+struct FootprintMargin
+{
+  double lon;
+  double lat;
+};
+
+FootprintMargin calcFootprintMargin(
+  const geometry_msgs::msg::PoseWithCovariance & covariance, const double scale)
+{
+  const auto Cov_in_map = covariance.covariance;
+  Eigen::Matrix2d Cov_xy_map;
+  Cov_xy_map << Cov_in_map[0 * 6 + 0], Cov_in_map[0 * 6 + 1], Cov_in_map[1 * 6 + 0],
+    Cov_in_map[1 * 6 + 1];
+
+  const double yaw_vehicle = tf2::getYaw(covariance.pose.orientation);
+
+  // To get a position in a transformed coordinate, rotate the inverse direction
+  Eigen::Matrix2d R_map2vehicle;
+  R_map2vehicle << std::cos(-yaw_vehicle), -std::sin(-yaw_vehicle), std::sin(-yaw_vehicle),
+    std::cos(-yaw_vehicle);
+  // Rotate covariance E((X, Y)^t*(X, Y)) = E(R*(x,y)*(x,y)^t*R^t)
+  // when Rotate point (X, Y)^t= R*(x, y)^t.
+  const Eigen::Matrix2d Cov_xy_vehicle = R_map2vehicle * Cov_xy_map * R_map2vehicle.transpose();
+
+  // The longitudinal/lateral length is represented
+  // in Cov_xy_vehicle(0,0), Cov_xy_vehicle(1,1) respectively.
+  return FootprintMargin{Cov_xy_vehicle(0, 0) * scale, Cov_xy_vehicle(1, 1) * scale};
+}
+}  // namespace
+
 namespace autoware::lane_departure_checker::utils
 {
 TrajectoryPoints cutTrajectory(const TrajectoryPoints & trajectory, const double length)
@@ -90,5 +122,43 @@ TrajectoryPoints resampleTrajectory(const Trajectory & trajectory, const double 
   resampled.push_back(trajectory.points.back());
 
   return resampled;
+}
+
+std::vector<LinearRing2d> createVehicleFootprints(
+  const geometry_msgs::msg::PoseWithCovariance & covariance, const TrajectoryPoints & trajectory,
+  const autoware::vehicle_info_utils::VehicleInfo & vehicle_info,
+  const double footprint_margin_scale)
+{
+  // Calculate longitudinal and lateral margin based on covariance
+  const auto margin = calcFootprintMargin(covariance, footprint_margin_scale);
+
+  // Create vehicle footprint in base_link coordinate
+  const auto local_vehicle_footprint = vehicle_info.createFootprint(margin.lat, margin.lon);
+
+  // Create vehicle footprint on each TrajectoryPoint
+  std::vector<LinearRing2d> vehicle_footprints;
+  for (const auto & p : trajectory) {
+    vehicle_footprints.push_back(
+      transformVector(local_vehicle_footprint, autoware::universe_utils::pose2transform(p.pose)));
+  }
+
+  return vehicle_footprints;
+}
+
+std::vector<LinearRing2d> createVehicleFootprints(
+  const PathWithLaneId & path, const autoware::vehicle_info_utils::VehicleInfo & vehicle_info,
+  const double footprint_extra_margin)
+{
+  // Create vehicle footprint in base_link coordinate
+  const auto local_vehicle_footprint = vehicle_info.createFootprint(footprint_extra_margin);
+
+  // Create vehicle footprint on each Path point
+  std::vector<LinearRing2d> vehicle_footprints;
+  for (const auto & p : path.points) {
+    vehicle_footprints.push_back(transformVector(
+      local_vehicle_footprint, autoware::universe_utils::pose2transform(p.point.pose)));
+  }
+
+  return vehicle_footprints;
 }
 }  // namespace autoware::lane_departure_checker::utils
